chore(cosyvoice3): remove development audit tools
Keep the onboarding PR focused on build/runtime support and required tests.
Remove the standalone diagnostics and validation command-line modules,
along with tests specific to those removed tools. Retain only the reference
helpers used by correctness tests under the family tests directory.

Do not change runtime code or numerical thresholds. The retained family
suite passed 182 tests with the previously validated checkpoint-build E2E
deselected. Preserve the development tools on the original feature branch.
