--- families/cosyvoice3/diagnostics/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""CosyVoice3 diagnostics; not imported by the inference runtime."""
