--- families/cosyvoice3/validation/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""CosyVoice3 validation; not imported by the inference runtime."""
